scripts/py-scripts/requirements.txt
google-cloud-storage
google-cloud-aiplatform
PyPDF2
